은형님 리덕스전에 커밋
